Fixxed a lot of stuff

+++ src/Main/java/Tickets/OtherTicket.java
@@ -2,12 +2,10 @@ package Tickets;
 
 import User.User;
 
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 
 public class OtherTicket extends Ticket {
-    public OtherTicket(String description, HashMap<User, Integer> users, int value, String purchaseDate, boolean splitEvenly){
+    public OtherTicket(String description, HashMap<User, Double> users, double value, String purchaseDate, boolean splitEvenly){
         super(description, users, value, purchaseDate, splitEvenly);
     }
 }
